Enable STS profile
We cannot have a mock profile which returns mock(DataSource.class), because then Hibernate/Jpa will raise a
NullPointerException basically saying it cannot build a proper connection to the database.
> [PersistenceUnit: default] Unable to build Hibernate SessionFactory; nested exception is java.lang.NullPointerException

The relevant tests are properly mocked so they don't need the database at all. However, as we are inside
of the framework we still need an environment with proper database to make tests pass.